Tcl cmds that take the "set' argument now document themselves in the help message for the command.
But to get this, you have to NOT override help_string(), which several
commands with bind_*() are still doing. Will fix next.

More manual work:
	* first two manpages, and system to build them into text
	* updated todo to ask for help finding a man to HTML formatter
	* more updates to sync up with big code changes

Added a -h option to dtnrecv, because it was just too simple with
no optional args...

Fixed a typo bug in UDP Convergence layer making it impossible to
bind to 0.0.0.0.