Update from MITRE provided by Scott, Keith L. <kscott@mitre.org>
The high-level list of modifications to the DTN2 implementation is:

o Added the ability for the DTN2 router component to modify locally-generated bundles before
they are delivered to local registrations (destination applications that use the same DTN2
daemon as the sender).

o Added support for transactionalized access to the DTN2 database used to store metadata
(bundle headers and application registration information).

o Added support for ODBC databases as a peer to the other database types provided with
the base open-source distribution (e.g. file-based, BerkeleyDB).

o Added support for flexible and efficient synchronization to disk, allowing the
implementation to run in ACI (atomicity, consistency, isolation) mode with
synchronization to disk (ACID mode) with every Nth transaction.

o For the ODBC database implementation, added support for loading external (non-DTN2)
schema for on database initialization.

o Ensured that all required data structures are in fact synchronized to disk.

o Added the ability for applications to acknowledge receipt of bundles to the DTN2 stack.

o Added support for a class of DTN router-private (or any DTN component-private) event types.

o Augmented the DTN2 application registration mechanism to allow applications to provide their
own unique token on registration (to support applications with no local storage).

o Modified the oasys locking and debugging infrastructure to allow non-oasys threads to take
locks without hanging the debugging machinery.

o Added support for nested event processing.

o Changed DTN2's DAEMON_SHUTDOWN event to NOT be a daemon-only event.

o Fixed a bug in the bundle find functions to prevent problems searching for bundles that are in
the process of being deleted.

o Fixed a bug in the bundle metadata processing to ensure that metadata blocks for locally-
generated bundles are correctly restored when the DTN2 daemon is restarted.

o Fixed a bug in the DTN2 IP Discovery module that was setting the socket options incorrectly
(more specifically, setting the correct option for the wrong socket) and generating copious
warnings in the log file.



This public release request pertains to enhancements to the DTN2 Reference Implementation [DTN2] of
the Bundle Protocol [RFC5050]. The bulk of the changes make the implementation more robust against
unexpected application termination (e.g. due to unexpected loss of power) and enable better
coordination between DTN2 routing components and the rest of the DTN2 system. Additional
capabilities were also added to give applications more control over what data is delivered to them
on reconnecting to the DTN router. Finally, several bugs in the DTN2 code were also addressed.


This work was supported by MITRE project #0711A790-SS. The government POC for this work is Dr. James
Richardson, PEO Integration, PdM SWI (james.p.richardson1@us.army.mil).