Patch provided by Elwyn Davies <elwynd@folly.org.uk> to address bug that could well result in a crash.
The interface signatures of begin/endTransaction in the
BerkeleyDBStore.h do not match those in the DurableStoreImpl.h whereas
they do in ODBCstore.h.

The result is that the transaction id is not set up
correctly (it should be an 'out' parameter to beginTransaction but the
default code just leaves the value in the caller untouched).