Patch from Elwyn Davies <davieseb@scss.tcd.ie> to fix the four major issues with ODBC database access:
1. Non-functionality of tables using string keys (links and prophet).

2. Failure to clear out old bound parameters before reusing hstmt
handles in ODBC.

3. Wrongly coded serialize_all_ scope locks, leading to interleaved
accesses by multiple threads to the ODBC hstmt handles.

4. Timeout of idle database connections.


=======================================================================

Changes to fix issues in the ODBC storage sub-system:
1. Non-functionality of tables using string keys (links and prophet).
2. Failure to clear out old bound parameters before reusing hstmt
handles in ODBC.
3. Wrongly coded serialize_all_ scope locks, leading to interleaved
accesses by multiple threads to the ODBC hstmt handles.
4. Timeout of idle database connections.
5. Removal of redundant (alleged) deadlock timer (deadlock is not
handled by timers in ODBC).

Additional minor fixes to various storage modules around transaction
(non-)support.

1. Non-functionality of tables using string keys (links and prophet).

Major reworking of ODBCStore to correctly handle keys other than unsigned
integers in a general fashion.  This came to light as a result of trying to
store link information which is keyed by the string name of the link.  As a
result all the special case code for the globals table has been able to be
removed.  The keys are now either VARBINARY for strings or fixed size BINARY
for integers.  The selection is controlled by the key_size_ which is set by a
routine in the key TypeShims.

2. Failure to clear out old bound parameters before reusing hstmt handles in ODBC.

It was also discovered that there was a
potential problem with old bound parameters.  ODBC statement handles rightly
or wrongly have a persistent memory for both bound parameters and bound
columns.  This is an issue if a use of the handle binds several parameters and
doesn't reset them before leaving the routine.  A later use of the same handle
will potentially barf if one or more of these old bound parameters is not
superseded by a new binding.  This is because they the binding may contain
stale addresses that now point to arbitrary bits of memory.  If SQLExecute
returns SQL_NEED_DATA (code 99) unexpectedly, this is a possible cause of the
problem. the ODBC driver looks at *all* the bound parameters whether they are
used by the prepared SQL statement or are just left around.  The fix is to
call SQLFreeStmt with the SQL_RESET_PARAMS option as well as with SQL_CLOSE
(which releases old bound columns) before the statement handle is reused.

3. Wrongly coded serialize_all_ scope locks, leading to interleaved accesses by multiple threads to the ODBC hstmt handles.

The alleged complete serialization of all database accesses in ODBCStore.cc
selected when serialize_all_ is set to true (default, no option to change via
command interface) turned out to be incorrectly implemented.  The code snippet

    if ( store_->serialize_all_&& !(store_->serialization_lock_.is_locked_by_me()) ) {
            ScopeLock sl(&store_->serialization_lock_, "Access by xxx()");
    }

which appears in several routines is effectively a no-op because the
ScopeLock sl is released at the trailing } in the if statement.  This leads to
occasional PANICs with incomprehensible 1064 (SQL parser error) codes from the
MySQL database quoting garbage strings allegedly seen in the submitted SQL.
This occurs when two threads in the daemon are trying to do database updates
at the same time (typically one of them will be the global store updating the
next_bundleid_ which is run on a convergence layer thread or the API server
thread when a new bundle is received whereas most otheraccesses run on the
main event handler thread).

The problem is fixed by ensuring that the various routines in ODBCStore.cc
really are properly serialized.  This is achived by replacing the above code
snippet with:

    ScopeLockIf sl(&store_->serialization_lock_,
                   "Access by xxx",
                   store_->serialize_all_);

wherever it occurs.  Also serialization is now applied to the
begin_transaction, end_transaction and del routines that also access the
database.

4. Timeout of idle database connections.

Add keep alive timer for MySQL connections (otherwise they time out after 8
hours by default - adjustable in MySQL configuration). Issues a trivial
request for server configuration information at periodic intervals (default
10 minutes - could be possibly set to 1 day with the MySQL configuration
set to remove idle sessions after (say) 26 hours.

5. Removal of redundant (alleged) deadlock timer.

Deadlock is not handled by timers in ODBC. Deadlock timer has no function
with ODBC (rather than BerkeleyDB) as deadlocks are detected by server
and reported as errors.

Clean up of ODBCStore:
 - Rename various class variables to end in _.
 - Removed all code that was previously commented out plus realKey() method.
 - Removed unused db_type from get_table and ODBCDBTable constructor/private
   data - this is a hangover from BerkeleyDB which is not relevant to ODBC/SQL -
   there is only one sort of SQL database from our point of view (SQLite
   doesn't have any options and it is really only sensible to use the InnoDB
   engine in MySQL as the MyISAM engine doesn't do row locking. InnoDB should
   be configured globally in the MySQL my.cnf file.)
 - Removed unnecessary base_dbenv_ item - made dbenv_ a structure rather
   than a pointer.
 - Removed cur_table from ODBCdbenv - the structure only applies globally and
   the concept of a current table is just confusing - in practice it was only
   used inside get_table.

Additionally, a number of minor changes have been applied to document that
- the MemoryStore is not transactional
- the FileSystemStore is not really transactional but end_transaction has been
  implemented to call a (new) routine in OpenFDCache.h that applies fsync() to
  all currently open file descriptors in the FileSystemStore fd_cache_,
  thereby flushing any pending updates to disk.
- to change the level of some messages associated with transactions to debug
  rather than info to make the normal operation less verbose
- make the default begin_transaction and end_transaction return DS_OK rather
  than DS_ERR - this was confusing and turning on debug logging gives all the
  necessary information should it be needed.

================================================================================